+++ trellis/modules/sparse/transformer/modulated.py
@@ -140,23 +140,29 @@ class ModulatedSparseTransformerCrossBlock(nn.Module):
             )
 
     def _forward(self, x: SparseTensor, mod: torch.Tensor, context: torch.Tensor) -> SparseTensor:
+        # pdb.set_trace()
+
         if self.share_mod:
             shift_msa, scale_msa, gate_msa, shift_mlp, scale_mlp, gate_mlp = mod.chunk(6, dim=1)
         else:
             shift_msa, scale_msa, gate_msa, shift_mlp, scale_mlp, gate_mlp = self.adaLN_modulation(mod).chunk(6, dim=1)
-        h = x.replace(self.norm1(x.feats))
+
+        h = x.replace(self.norm1.float()(x.feats))
         h = h * (1 + scale_msa) + shift_msa
         h = self.self_attn(h)
         h = h * gate_msa
         x = x + h
-        h = x.replace(self.norm2(x.feats))
+        h = x.replace(self.norm2.float()(x.feats))
         h = self.cross_attn(h, context)
         x = x + h
-        h = x.replace(self.norm3(x.feats))
+        h = x.replace(self.norm3.float()(x.feats))
         h = h * (1 + scale_mlp) + shift_mlp
         h = self.mlp(h)
         h = h * gate_mlp
         x = x + h
+
+        # pdb.set_trace()
+
         return x
 
     def forward(self, x: SparseTensor, mod: torch.Tensor, context: torch.Tensor) -> SparseTensor:
